Phase_3/UserInterface/requirements.txt
psycopg2-binary==2.9.3
tabulate==0.8.9
